content/TryYours-Virtual-Try-On/requirements.txt
Flask
numpy
pandas
pillow
python-dotenv
tensorflow
scikit-learn
flask
flask-cors
opencv-python-headless
gunicorn
tensorboardX
torchgeometry
flask-ngrok
iglovikov_helper_functions
cloths_segmentation
albumentations
scipy==1.8.0